move tuf pkg to nest under autoupdate
