+++ pkg/autoupdate/tuf/autoupdate_test.go
